c2

+++ src/main/java/vladislavmaltsev/paymenttaskapi/repository/UserPaymentRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserPaymentRepository extends JpaRepository<User, Integer> {
-        Optional<User> findById(int id);
+        Optional<User> findById(long id);
         Optional<User> findByName(String name);
 }
